fix slow pirate movement in between starfish
We have some extraneous moveTo actions from the center of a tile to the exact
actor (starfish) position, for each starfish in the route. Right now these are
superfluous, as the starfish are always placed in the center of tiles. So
skipping these actions results in faster movement. The only one of these
adjusting moves that is needed is at the end of the route.