Allow for multiple treasures/goals.
Level is complete only if all treasures are acquired.
There is currently a problem where placing a starfish on an existing route (i.e.
backtracking or making circle-routes) doe not work - maybe because of the
foottrail - not sure. As a temporary workaround, change level 3 to work without
a backtrack / loop of the hero route.